Ticket cancellation validation added for one hour

+++ cusr/src/main/java/edu/sjsu/cmpe275/cusr/service/TicketService.java
@@ -1,6 +1,12 @@
 package edu.sjsu.cmpe275.cusr.service;
 
+import java.text.DateFormat;
+import java.text.ParseException;
+import java.text.SimpleDateFormat;
+import java.time.Duration;
+import java.time.LocalDateTime;
 import java.util.ArrayList;
+import java.util.Date;
 import java.util.List;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Component;
@@ -35,26 +41,61 @@ public class TicketService {
 		return ticketRepository.save(ticket);
 	}
 	
-	public Ticket cancelTicket(Ticket ticket)
+	public boolean cancelTicket(Ticket ticket)
 	{
 		ArrayList<JourneyDetails> cancelledJourney = journeyRepository.findManyByTicket(ticket.getTicketId());
+		boolean isTicketAllowedToCancel = false;
+		
 		for(JourneyDetails jd: cancelledJourney){
-			journeyRepository.delete(jd);
+			
+			String journeyDate = jd.getJourneyDate();
+			String deptTime = jd.getDepartureTime();
+			
+			String value = journeyDate+"/"+deptTime;
+			
+			DateFormat dateFormat = new SimpleDateFormat("yyyy-MM-dd/HHmm");
+			Date ticketDate = null;
+			try {
+				ticketDate = dateFormat.parse(value);
+			} catch (ParseException e) {
+				e.printStackTrace();
+			}
+			
+			LocalDateTime now = LocalDateTime.now();
+			LocalDateTime ticketThen = LocalDateTime.of(ticketDate.getYear()+1900, 
+					ticketDate.getMonth()+1, 
+					ticketDate.getDate(), ticketDate.getHours(), ticketDate.getMinutes());
+			
+			Duration duration = Duration.between(ticketThen, now);
+			long diff = Math.abs(duration.toMinutes());
+			
+			if(diff >= 60){
+				isTicketAllowedToCancel = true;
+			} else{
+				isTicketAllowedToCancel = false;
+				break;
+			}
 		}
-
-		Transaction cancelledTicketTransaction = transactionRepository.findOneByTicket(ticket.getTicketId());
-		transactionRepository.delete(cancelledTicketTransaction);
 		
-		ArrayList<Passenger> cancelledTicketPassengers = passengerRepository.findManyByTicket(ticket.getTicketId());
-		for(Passenger passenger : cancelledTicketPassengers)
-		{
-			passengerRepository.delete(passenger);
+		if(isTicketAllowedToCancel){
+			for(JourneyDetails jd: cancelledJourney){
+				journeyRepository.delete(jd);
+			}
+	
+			Transaction cancelledTicketTransaction = transactionRepository.findOneByTicket(ticket.getTicketId());
+			transactionRepository.delete(cancelledTicketTransaction);
+			
+			ArrayList<Passenger> cancelledTicketPassengers = passengerRepository.findManyByTicket(ticket.getTicketId());
+			for(Passenger passenger : cancelledTicketPassengers)
+			{
+				passengerRepository.delete(passenger);
+			}
+			
+			ticket.setCancelled(true);
+			ticketRepository.save(ticket);
 		}
 		
-		ticket.setCancelled(true);
-		ticketRepository.save(ticket);
-		
-		return ticket;
+		return isTicketAllowedToCancel;
 	}
 	
 	public Ticket getTicketById(long id)
